make regiment.id a u32, not u16
Blueprint `Node.regiment_id` is a u32. Previously `Regiment.id` was a
u16 and there were two unknown bytes after it. Those unknown bytes are
always 0 and so it's highly likely that `Regiment.id` was meant to be a
u32. The way to prove this in game would be to write some non-zero bytes
to the unknown slots and see if anything happens.