incorporate source audio file sample rate in random_playback_rate calc (#251)
If you listen to APPEAR01.WAV alone, it doesn't appear to be a sound
used in game. However, it's actually used for the main menu button hover
sound and the toggling magic items panel sound. It's just that it's
playback rate was calculated incorrectly by not taking into account the
sample rate of the source. This change fixes that. It's possible that
the frequency deviation formula isn't 100% accurate to the original, but
at the very least it doesn't appear to sound worse than before this
change, and the toggle magic items panel sound is now correct.